feat: Fully refactor the chat system to use form actions
